Update README.md
I got rid of the reference to the out-of-date README template that is housed on the LIS filesystem but belongs to a repo under LegumeFederationDataStore which I think should be blown away as it's super confusing and contains inaccurate, out-of-date files. I'll add a new README.collection.yml right here as a template, and this repo could be cloned onto the LIS filesystem if we like. @cann0010 @adf-ncgr