Update qtlmrk.md
Caught a mistake in this spec: sometimes you have QTL-marker relationships but no QTL spans on a LG (which is fine). But in this case the QTL isn't listed in the qtl.tsv file with its trait; so we need to include the trait in the qtlmrk.tsv file as well, otherwise we get a stranded QTL not linked to a trait (and OBO terms).