Update qtlmrk.md
Another tweak. Often, a marker is positioned on multiple linkage groups. But in a particular study, the QTL-marker relationship is found on a particular linkage group. So let's optionally list that linkage group in the qtlmrk.tsv file so we know what the experimenters were doing. This can be interesting: for 10/11 studies in SoyBase, Satt547 is positioned on GmComposite2003_J; but for one, it was positioned on GmComposite2003_ZZ, whatever that is (which could be a typo).

A good example of a marker mapped to several genetic maps:
```
A082_1  GmComposite1999_A1      95.40
A082_1  GmComposite2003_A1      102.30
A082_1  GmRFLP-KGL_L30             0.00
A082_1  GmRFLP-USDAARS-RCS_A    108.50
A082_1  GmUSDA1997_A1_1997      128.50
```