Create seqid_map.md
I've added the third column in this new seqid_map.tsv file spec, @adf-ncgr and @cann0010 . Note that I say "supercontig," since "scaffold" is not a core SO term, it's a synonym that redirects to "supercontig". (Which is why it's been Supercontig in the mines since day 1.) "contig" is a core SO term, with "supercontig" as a parent.